+++ NEWS.md
@@ -1,7 +1,9 @@
 # sfdep (development version)
 
+- set operations for neighbor lists `nb_union()`, `nb_intersect()`, `nb_setdiff()`
 - `nb_as_matrix()` and `wt_as_matrix()` to convert neighbor or weights lists to matrix representation
 - `st_block_nb()` is introduced to create neighbor contiguity based on spatial regimes
+  - inspired by https://pysal.org/libpysal/generated/libpysal.weights.block_weights.html
 - `critical_threshold()` gains argument `k` to identify minimum number of neighbors when calculating threshold
 - `pct_nonzero()` calculates the percent of non-zero neighbors
 - `szero()` to calculate the global sum of weights